Break init.lua into modules + move to colorschemes
