+++ nvim/lua/ui.lua
@@ -1,0 +1,2 @@
+-- @see https://github.com/nvim-tree/nvim-web-devicons
+require('nvim-web-devicons').setup()
